+++ Runtime/Serie/SeriesHelper.cs
@@ -346,7 +346,7 @@ namespace XCharts.Runtime
                     if ((isPolar && serie.polarIndex != axisIndex) ||
                         (!isPolar && serie.yAxisIndex != axisIndex) ||
                         !serie.show) continue;
-                    var updateDuration = serie.animation.enable?serie.animation.dataChangeDuration : 0;
+                    var updateDuration = serie.animation.enable ? serie.animation.dataChangeDuration : 0;
                     var unscaledTime = serie.animation.unscaledTime;
                     if (isPercentStack && SeriesHelper.IsPercentStack<Bar>(series, serie.serieName))
                     {
@@ -355,7 +355,7 @@ namespace XCharts.Runtime
                     }
                     else
                     {
-                        var showData = serie.GetDataList(filterByDataZoom?chart.GetXDataZoomOfSerie(serie) : null);
+                        var showData = serie.GetDataList(filterByDataZoom ? chart.GetXDataZoomOfSerie(serie) : null);
                         if (serie is Candlestick || serie is SimplifiedCandlestick)
                         {
                             foreach (var data in showData)
@@ -371,9 +371,9 @@ namespace XCharts.Runtime
                             var performanceMode = serie.IsPerformanceMode();
                             foreach (var data in showData)
                             {
-                                var currData = performanceMode? data.GetData(yValue?1 : 0, inverse):
+                                var currData = performanceMode ? data.GetData(yValue ? 1 : 0, inverse) :
                                     data.GetCurrData(yValue ? 1 : 0, updateDuration, unscaledTime, inverse);
-                                if (!serie.IsIgnoreValue(currData))
+                                if (!serie.IsIgnoreValue(data, currData))
                                 {
                                     if (currData > max) max = currData;
                                     if (currData < min) min = currData;
@@ -395,7 +395,7 @@ namespace XCharts.Runtime
                         if ((isPolar && serie.polarIndex != axisIndex) ||
                             (!isPolar && serie.yAxisIndex != axisIndex) ||
                             !serie.show) continue;
-                        var showData = serie.GetDataList(filterByDataZoom?chart.GetXDataZoomOfSerie(serie) : null);
+                        var showData = serie.GetDataList(filterByDataZoom ? chart.GetXDataZoomOfSerie(serie) : null);
                         if (SeriesHelper.IsPercentStack<Bar>(series, serie.stack))
                         {
                             for (int j = 0; j < showData.Count; j++)
@@ -419,7 +419,7 @@ namespace XCharts.Runtime
                                     currData = yValue ? showData[j].GetData(1) : showData[j].GetData(0);
                                 }
                                 if (inverse) currData = -currData;
-                                if (!serie.IsIgnoreValue(currData))
+                                if (!serie.IsIgnoreValue(showData[j], currData))
                                     _serieTotalValueForMinMax[j] = _serieTotalValueForMinMax[j] + currData;
                             }
                         }